get modules working while using live server
