+++ index.html
@@ -27,7 +27,6 @@
       </svg>
     </div>
     <script src="https://d3js.org/d3.v6.min.js"></script>
-    <script src="main.js"></script>
-    <script src="undo-redo.js"></script>
+    <script type="module" src="main.js"></script>
   </body>
 </html>
